changed icons

+++ shared/src/commonMain/kotlin/org/example/project/data/api/auth_api/model/CreateCartRequestBody.kt
@@ -11,6 +11,6 @@ class CreateCartRequestBody(
     val deliveryType: DeliveryType,
     val deliveryAddress: AddressEntity?,
     val departmentId: Int,
-    val deliveryPrice: Double,
-    val freeDeliveryPrice: Double?
+    val deliveryPrice: Long,
+    val freeDeliveryPrice: Long?
 )
