+++ shared/src/commonMain/kotlin/org/example/project/data/entities/CartItemEntity.kt
@@ -7,5 +7,5 @@ data class CartItemEntity(
     val productId: Long,
     val title: String,
     val quantity: Int,
-    val price: Float,
+    val price: Long,
 )
