+++ shared/src/commonMain/kotlin/org/example/project/domain/models/CartItemModel.kt
@@ -4,5 +4,5 @@ data class CartItemModel(
     val productId: Long,
     val title: String,
     val quantity: Int,
-    val price: Float,
+    val price: Long,
 )
